% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_qtr.R
\name{date_to_qtr}
\alias{date_to_qtr}
\title{date_to_qtr}
\usage{
date_to_qtr(date, type = "FYQ")
}
\arguments{
\item{type}{one of 3 options:

"FYQ" returns the financial year and quarter
"Q" returns the quarter only
"FY" returns the financial year only}
}
\description{
given a date, this should return the NHS financial year, quarter, or financial year and quarter
}
\examples{
date_to_qtr(as.Date('2022-04-01' ,format = "\%Y-\%m-\%d"))
date_to_qtr(as.Date('2022-04-01' ,format = "\%Y-\%m-\%d"), type="Q")
date_to_qtr(as.Date('2022-04-01' ,format = "\%Y-\%m-\%d"), type="FY")
date_to_qtr(as.Date('04-01-2022' ,format = "\%d-\%m-\%Y"),type="FYQ")
date_to_qtr(as.Date('04-01-2022' ,format = "\%d-\%m-\%Y"),type="FYQ2")
}
