% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_bands.R
\name{generate_age_bands}
\alias{generate_age_bands}
\title{Suggest Code for Creating Age Bands}
\usage{
generate_age_bands(col = "Age", age_bands = NULL, common_age_bands = 999)
}
\arguments{
\item{col}{Name of age column to transform.}

\item{age_bands}{This should be a vector of age bands containing numeric values such as c("0-14","15-65","65+"), where the
order of the vector specifies the order with which the age_band appears within the returned dplyr::case_when statement.

This parameter can be left blank if common_age_bands has been specified as 1,2 or 3.
In these cases the age bands corresponding to these values (as specified in common_age_bands parameter) will be used.}

\item{common_age_bands}{When this is null or outside the specified values below, the function
will return dplyr code corresponding to the age bands specified in the parameter age_bands
If, instead, a numeric value 1,2 or 3 is inputted for this parameter, the following age bands below will be used instead:

1 = c("0-4","5-9","10-14","15-19","20-24","25-29","30-34","35-39","40-44","45-49",
"50-54","55-59","60-64","65+")

2 = c("0-4","5-9","10-14","15-19","20-24","25-29","30-34","35-39","40-44","45-49",
"50-54","55-59","60-64","65-69","70-74","75-79","80-84","85-89","90-94","95-99","100+")

3 = c("0-9","10-19","20-29","30-39","40-49","50-59","60-69","70-79","80-89","90-99","100+").}
}
\value{
A statement printed in the console, giving the suggested dplyr code to create age bands from a
age column
}
\description{
Prints out suggested dplyr code to generate age bands. This is intended to serve as a guide only
and therefore code generated should be reviewed and modified as necessary.
}
\examples{


data <- data.frame(
  Age = c(
    5L, 110L, 87L, 108L, 57L, 168L, 133L, 62L,
    125L, 140L, 38L, 168L, 159L, 155L, 98L, 37L, 24L, 48L, 141L,
    34L, 14L, 129L, 70L, 121L, 102L, 8L, 113L, 113L, 3L, 71L, 111L,
    95L, 41L, 26L, 170L, 160L, 73L, 76L, 84L, 106L, 73L, 1L, 58L,
    190L, 22L, 118L, 52L, 79L
  ),
  ID = c(
    2251558L, 2540739L, 2121923L, 268210L,
    1695141L, 103196L, 485232L, 2431101L, 1565663L, 1687369L, 762799L,
    293488L, 1476981L, 1764643L, 2601051L, 463056L, 867640L,
    426493L, 1424699L, 1166167L, 1314136L, 2079697L, 1230127L, 1055453L,
    2529562L, 909298L, 831791L, 2346930L, 2341395L, 573434L,
    1669739L, 1501823L, 590632L, 186521L, 495621L, 1007120L, 1404587L,
    626912L, 2593705L, 2148515L, 1051717L, 1540477L, 1337032L,
    1694720L, 101417L, 1592089L, 2065567L, 93434L
  ),
  Value = c(
    3L, 7L, 10L, 10L, 6L, 10L, 8L, 4L, 1L, 3L,
    10L, 4L, 2L, 2L, 8L, 7L, 1L, 7L, 1L, 10L, 3L, 9L, 6L, 8L, 4L,
    5L, 1L, 1L, 10L, 2L, 3L, 8L, 4L, 8L, 5L, 4L, 5L, 9L, 8L, 4L, 8L,
    7L, 4L, 2L, 7L, 6L, 8L, 2L
  )
)

age_bands2 <- c("<5", "5-9", "10-15", "20-99", "100-119", "120+")

generate_age_bands("Age", age_bands2)
generate_age_bands("Age", age_bands2, 2)

}
\seealso{
\code{\link[=order_age_bands]{order_age_bands()}} for suggesting code to order age bands
}
