% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_bands.R
\name{order_age_bands}
\alias{order_age_bands}
\title{Suggest Code for Ordering of Age Bands}
\usage{
order_age_bands(y)
}
\arguments{
\item{y}{This parameter should be a vector or data frame column of age bands.}
}
\value{
A vector of the unique age band values reordered
}
\description{
Returns a suggested ordering of age bands. This is intended to serve as a guide only
and therefore code generated should be reviewed and modified as necessary.

The function will output an vector containing the sugggested ordering, and also
print out the code needed to create this vector, to support any modification required.

The function is intended for use in creating visualisations ordered by age groups.
}
\examples{


data3 <- data.frame(
  stringsAsFactors = FALSE,
  Age = c(
    5L, 110L,
    87L, 108L, 57L, 168L,
    133L, 62L, 125L,
    140L, 38L, 168L, 159L,
    155L, 98L, 37L,
    24L, 48L, 141L, 34L,
    14L, 129L, 70L, 121L,
    102L, 8L, 113L,
    113L, 3L, 71L, 111L,
    95L, 41L, 26L, 170L,
    160L, 73L, 76L, 84L,
    106L, 73L, 1L, 58L,
    190L, 22L, 118L, 52L,
    79L
  ),
  ID = c(
    2251558L,
    2540739L, 2121923L,
    268210L, 1695141L,
    103196L, 485232L,
    2431101L, 1565663L,
    1687369L, 762799L,
    293488L, 1476981L,
    1764643L, 2601051L,
    463056L, 867640L, 426493L,
    1424699L, 1166167L,
    1314136L, 2079697L,
    1230127L, 1055453L,
    2529562L, 909298L,
    831791L, 2346930L,
    2341395L, 573434L,
    1669739L, 1501823L,
    590632L, 186521L,
    495621L, 1007120L,
    1404587L, 626912L,
    2593705L, 2148515L,
    1051717L, 1540477L,
    1337032L, 1694720L,
    101417L, 1592089L,
    2065567L, 93434L
  ),
  Value = c(
    3L, 7L,
    10L, 10L, 6L, 10L, 8L,
    4L, 1L, 3L, 10L, 4L,
    2L, 2L, 8L, 7L, 1L,
    7L, 1L, 10L, 3L,
    9L, 6L, 8L, 4L, 5L,
    1L, 1L, 10L, 2L, 3L,
    8L, 4L, 8L, 5L, 4L,
    5L, 9L, 8L, 4L, 8L,
    7L, 4L, 2L, 7L, 6L,
    8L, 2L
  ),
  age_group = c(
    "5-9",
    "100+", "85-89", "100+",
    "55-59", "100+",
    "100+", "60-64",
    "100+", "100+", "35-39",
    "100+", "100+",
    "100+", "95-99",
    "35-39", "20-24", "45-49",
    "100+", "30-34",
    "10-14", "100+",
    "70-74", "100+", "100+",
    "5 9", "100+",
    "100+", "<4", "70-74",
    "100+", "95-99",
    "40-44", "25-29", "100+",
    "100+", "70-74",
    "75-79", "80-84",
    "100+", "70-74", "<4",
    "55-59", "100+",
    "20-24", "100+",
    "50-54", "75-79"
  )
)

order_age_bands(data3$age_group)

data3$age_group <- factor(x = data3$age_group, levels = order_age_bands(data3$age_group), ordered = T)
ggplot2::ggplot(data3, ggplot2::aes(x = age_group)) +
  ggplot2::geom_bar(stat = "count")
}
\seealso{
\code{\link[=generate_age_bands]{generate_age_bands()}} for suggesting code to create age bands from an age vector
}
